requirements.txt
gradio
requests
pandas
python-dotenv
langchain-core
langchain-community
langchain-openai
langgraph
pdfplumber
playwright
beautifulsoup4
Pillow
yt-dlp
youtube-transcript-api
wolframalpha
pytesseract==0.3.13
pydantic==2.11.3
pypdfium2==4.30.1
